airflow-poc/requirements.txt
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-amazon==8.7.1
psycopg2-binary==2.9.9
pandas==2.0.3
pyarrow==13.0.0
boto3==1.28.85
python-dotenv==1.0.0
sqlalchemy==1.4.53
fastparquet==2023.8.0
numpy==1.24.3
faker==19.6.2
tqdm==4.66.1
